requirements.txt
flask==2.3.3
pyodbc==5.1.0
python-dotenv==1.0.1
gunicorn==22.0.0